setting debug macro and more config

+++ src/SMEIoT.Web/ServerSetup.cs
@@ -1,21 +1,26 @@
 using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Connections;
+using Microsoft.Extensions.Configuration;
 using SMEIoT.Infrastructure.Mosquitto;
 using System.IO;
+using System;
 
 namespace SMEIoT.Web
 {
   public static class ServerSetup
   {
-    public static void ConfigureKestrel(KestrelServerOptions options)
+    public static void ConfigureKestrel(WebHostBuilderContext context, KestrelServerOptions options)
     {
-      var unixSocket = "/var/SMEIoT/run/smeiot.auth.broker";
+      var unixSocket = context.Configuration.GetSection("SMEIoT")?.GetValue<string>("MosquittoLocalAuthenticationSocket");
+      if (unixSocket == null) {
+        return;
+      }
+
       if (File.Exists(unixSocket))
       {
         File.Delete(unixSocket);
       }
-      // TODO: don't use this address in dev. might be tricky to config
       options.ListenUnixSocket(unixSocket, builder =>
       {
         builder.Protocols = HttpProtocols.None;
